--- pyproject.toml
@@ -1,18 +1,0 @@
-[tool.poetry]
-name = "tainanpublicwitnessing"
-version = "0.1.0"
-description = ""
-authors = ["xellosiris <xellosiris@gmail.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.10"
-firebase-admin = "^5.4.0"
-pandas = "^1.5.0"
-requests = "^2.28.1"
-
-[tool.poetry.dev-dependencies]
-ipykernel = "^6.16.0"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
