--- requirements.txt
@@ -1,34 +1,0 @@
-cachecontrol==0.12.12; python_version >= "3.6"
-cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
-certifi==2022.9.24; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy"
-charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0" and platform_python_implementation != "PyPy"
-firebase-admin==5.4.0; python_version >= "3.6"
-google-api-core==2.10.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
-google-api-python-client==2.63.0; python_version >= "3.7"
-google-auth-httplib2==0.1.0; python_version >= "3.7"
-google-auth==2.12.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0"
-google-cloud-core==2.3.2; platform_python_implementation != "PyPy" and python_version >= "3.7"
-google-cloud-firestore==2.7.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
-google-cloud-storage==2.5.0; python_version >= "3.7"
-google-crc32c==1.5.0; python_version >= "3.7"
-google-resumable-media==2.3.3; python_version >= "3.7"
-googleapis-common-protos==1.56.4; platform_python_implementation != "PyPy" and python_version >= "3.7"
-grpcio-status==1.49.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
-grpcio==1.49.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
-httplib2==0.20.4; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
-idna==3.4; python_version >= "3.7" and python_version < "4" and platform_python_implementation != "PyPy"
-msgpack==1.0.4; python_version >= "3.6"
-numpy==1.23.3; python_version >= "3.10"
-pandas==1.5.0; python_version >= "3.8"
-proto-plus==1.22.1; platform_python_implementation != "PyPy" and python_version >= "3.7"
-protobuf==4.21.6; platform_python_implementation != "PyPy" and python_version >= "3.7"
-pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0"
-pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0") or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
-pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7"
-python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
-pytz==2022.2.1; python_version >= "3.8"
-requests==2.28.1; python_version >= "3.7" and python_version < "4"
-rsa==4.9; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
-six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.8" and platform_python_implementation != "PyPy"
-uritemplate==4.1.1; python_version >= "3.7"
-urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" and platform_python_implementation != "PyPy"
