+++ src/app/_api/api.interface.ts
@@ -5,8 +5,6 @@ import { ShiftHours } from '../_interfaces/shift-hours.interface';
 import { Shift } from '../_interfaces/shift.interface';
 import { PersonalShifts } from '../_interfaces/personal-shifts.interface';
 import { Statistic } from '../_interfaces/statistic.interface';
-import { SiteShifts } from '../_interfaces/site-shifts.interface';
-import { UserSchedule } from '../_interfaces/user-schedule.interface';
 
 export interface ApiInterface {
   login: (uuid: string, password: string) => Promise<void>;
@@ -15,11 +13,7 @@ export interface ApiInterface {
   readUserKeys: () => Promise<UserKey[]>;
 
   readUser: (uuid: string) => Promise<User>;
-  /**
-   * create new user and return its uuid
-   * @return {string} new user uuid
-   */
-  createUser: (user: Omit<User, 'uuid' | 'activate'>) => Promise<string>;
+  createUser: (user: Omit<User, 'uuid' | 'activate'>) => Promise<void>;
   patchUser: (user: Omit<User, 'activate'>) => Promise<void>;
   updateUserActivation: (
     uuid: string,
@@ -57,12 +51,4 @@ export interface ApiInterface {
   readStatistic: (yearMonth: string, uuid: string) => Promise<Statistic>;
   createStatistic: (statistic: Statistic) => Promise<void>;
   updateStatistic: (statistic: Statistic) => Promise<void>;
-
-  readSiteShifts: () => Promise<SiteShifts[]>;
-
-  readUserSchedule: (userUuid: string) => Promise<UserSchedule>;
-  patchUserSchedule: (
-    userUuid: string,
-    data: Partial<UserSchedule>
-  ) => Promise<void>;
 }
