--- src/app/_directives/force-refresh.directive.spec.ts
@@ -1,8 +1,0 @@
-import { ForceRefreshDirective } from './force-refresh.directive';
-
-describe('AppRerenderDirective', () => {
-  it('should create an instance', () => {
-    const directive = new ForceRefreshDirective();
-    expect(directive).toBeTruthy();
-  });
-});
