--- src/app/_directives/force-refresh.directive.ts
@@ -1,16 +1,0 @@
-import { Directive, Input, TemplateRef, ViewContainerRef } from '@angular/core';
-
-@Directive({
-  selector: '[forceRefresh]',
-})
-export class ForceRefreshDirective {
-  @Input('forceRefresh') set forceRefresh(value: any) {
-    this.viewContainer.clear();
-    this.viewContainer.createEmbeddedView(this.templateRef);
-  }
-
-  constructor(
-    private templateRef: TemplateRef<unknown>,
-    private viewContainer: ViewContainerRef
-  ) {}
-}
