+++ src/app/_elements/shift-table/shift-card/shift-card.component.html
@@ -1,4 +1,4 @@
-<mat-card class="shift-card h-full box-border">
+<mat-card class="shift-card">
 
   <mat-card-header *ngIf="crew !== null">
