--- src/app/_interfaces/site-shifts.interface.ts
@@ -1,8 +1,0 @@
-export interface SiteShifts {
-  activate: boolean;
-  attendence: number;
-  shiftHoursUuid: string;
-  siteUuid: string;
-  uuid: string;
-  weekday: number;
-}
