--- src/app/_interfaces/user-schedule.interface.ts
@@ -1,20 +1,0 @@
-export interface UserSchedule {
-  availableHours: UserScheduleHours;
-  partnerUuid: string | null;
-  assign: boolean;
-  unavailableDates: string[];
-}
-
-export interface UserScheduleHours {
-  '0'?: UserScheduleDayData;
-  '1'?: UserScheduleDayData;
-  '2'?: UserScheduleDayData;
-  '3'?: UserScheduleDayData;
-  '4'?: UserScheduleDayData;
-  '5'?: UserScheduleDayData;
-  '6'?: UserScheduleDayData;
-}
-
-export interface UserScheduleDayData {
-  [ShiftHourUuid: string]: number;
-}
