--- src/app/_pipes/array-filter.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { ArrayFilterPipe } from './array-filter.pipe';
-
-describe('ArrayFilterPipe', () => {
-  it('create an instance', () => {
-    const pipe = new ArrayFilterPipe();
-    expect(pipe).toBeTruthy();
-  });
-});
