+++ src/app/_services/authority.service.ts
@@ -30,7 +30,6 @@ export class AuthorityService implements CanActivate {
     { url: 'users/:mode/:uuid?', permission: Permission.MANAGER },
     { url: `users/${Mode.CREATE}`, permission: Permission.ADMINISTRATOR },
     { url: 'users', permission: Permission.MANAGER },
-    { url: 'profile', permission: Permission.USER },
   ];
 
   constructor(
