+++ src/app/app.component.scss
@@ -33,13 +33,11 @@
       justify-content: center;
 
       div {
-        $padding: 16px;
-
         width: 100%;
+        max-width: $page-max-width;
         min-width: $page-min-width;
         height: calc(100% - 32px);
-        padding: 0 calc(max($padding, (100% - $page-max-width) / 2));
-        padding-bottom: $padding;
+        padding: 16px 16px 0 16px;
         overflow: auto;
       }
     }
