+++ src/app/app.component.ts
@@ -1,10 +1,4 @@
-import {
-  AfterViewInit,
-  Component,
-  OnDestroy,
-  OnInit,
-  ViewChild,
-} from '@angular/core';
+import { AfterViewInit, Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { MatSidenav } from '@angular/material/sidenav';
 import { FocusMonitor } from '@angular/cdk/a11y';
 import { TranslateService } from '@ngx-translate/core';
@@ -21,13 +15,14 @@ import { environment } from 'src/environments/environment';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
+
   @ViewChild('sidenav') sidenav!: MatSidenav;
   @ViewChild('menuButton') menuButton!: MatButton;
-
-  currentUsername$ = new BehaviorSubject<string | null>(null);
+  
+  currentUsername$ = new BehaviorSubject<string|null>(null);
   displayUsername$ = new BehaviorSubject<boolean>(true);
   languages: string[] = Object.values(Language);
   destroy$ = new Subject<void>();
@@ -38,49 +33,41 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
     private globalEventService: GlobalEventService,
     private translateService: TranslateService,
     private dateAdapter: DateAdapter<any>,
-    private focusMonitor: FocusMonitor
+    private focusMonitor: FocusMonitor,
   ) {}
 
   ngOnInit(): void {
-    fromEvent(window, 'resize')
-      .pipe(
-        takeUntil(this.destroy$),
-        map(() => window.innerWidth),
-        startWith(window.innerWidth),
-        map((width) => width > 475)
-      )
-      .subscribe((displayUsername) =>
-        this.displayUsername$.next(displayUsername)
-      );
+    fromEvent(window, 'resize').pipe(
+      takeUntil(this.destroy$),
+      map(() => window.innerWidth),
+      startWith(window.innerWidth),
+      map(width => width > 475),
+    ).subscribe(displayUsername => this.displayUsername$.next(displayUsername));
 
     this.translateService.setDefaultLang('zh');
     this.dateAdapter.setLocale(environment.MOMENT_LOCALES['zh']);
 
-    this.authorityService.currentUserUuid$
-      .pipe(
-        takeUntil(this.destroy$),
-        map((uuid) => {
-          if (uuid) {
-            return this.usersService.getUserByUuid(uuid).pipe(
-              filter((user) => user !== null),
-              map((user) => (user ? user.name : null))
-            );
-          } else {
-            return of(null);
-          }
-        }),
-        switchAll()
-      )
-      .subscribe((name) => this.currentUsername$.next(name));
+    this.authorityService.currentUserUuid$.pipe(
+      takeUntil(this.destroy$),
+      map(uuid => {
+        if (uuid) {
+          return this.usersService.getUserByUuid(uuid).pipe(filter(user => user !== null), map(user => user ? user.name : null));
+        } else {
+          return of(null);
+        }
+      }),
+      switchAll()
+    ).subscribe(name => this.currentUsername$.next(name));
+
+
   }
 
   ngAfterViewInit(): void {
-    this.globalEventService
-      .getGlobalEventById('ON_MENU_LINK_CLICK')
-      .pipe(takeUntil(this.destroy$))
-      .subscribe(() => {
-        this.sidenav.close();
-      });
+    this.globalEventService.getGlobalEventById('ON_MENU_LINK_CLICK').pipe(
+      takeUntil(this.destroy$)
+    ).subscribe(() => {
+      this.sidenav.close();
+    });
 
     this.focusMonitor.stopMonitoring(this.menuButton._elementRef.nativeElement);
   }
@@ -93,9 +80,9 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
   onLanguageButtonClick = (language: Language) => {
     this.translateService.use(language);
     this.dateAdapter.setLocale(environment.MOMENT_LOCALES['zh']);
-  };
+  }
 
   onMenuButtonClick = () => {
     this.sidenav.toggle();
-  };
+  }
 }
