+++ src/app/personal-shift/personal-shift.component.ts
@@ -3,10 +3,10 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { MAT_DATE_FORMATS } from '@angular/material/core';
 import { MatDatepicker } from '@angular/material/datepicker';
-import type { Moment } from 'moment';
 import * as moment from 'moment';
-import { forkJoin, Observable, of } from 'rxjs';
-import { filter, first, map, startWith, switchAll } from 'rxjs/operators';
+import { Moment } from 'moment';
+import { BehaviorSubject, combineLatest, forkJoin, Observable, of } from 'rxjs';
+import { filter, map, startWith, switchAll, first } from 'rxjs/operators';
 import { Shift } from '../_interfaces/shift.interface';
 import { AuthorityService } from '../_services/authority.service';
 import { PersonalShiftsService } from '../_services/personal-shifts.service';
