--- src/app/users/user/user-schedule/calendar-header/calendar-header.component.html
@@ -1,1 +1,0 @@
-<h4 class="mt-0">{{ "USERS.LABEL.UNAVILABLE_DATES" | translate }}</h4>
