--- src/app/users/user/user-schedule/hours-table/hours-table.component.html
@@ -1,30 +1,0 @@
-<table mat-table [dataSource]="data" class="w-full">
-  <ng-container matColumnDef="day">
-    <th mat-header-cell *matHeaderCellDef>
-      <h4 class="text-sm">{{ "USERS.LABEL.AVAILABLE_HOURS" | translate }}</h4>
-    </th>
-    <td mat-cell *matCellDef="let row">
-      {{ "GLOBAL.DAY." + weekdayNames[row.day] | translate }}
-    </td>
-  </ng-container>
-
-  <ng-container *ngFor="let hour of hours" [matColumnDef]="hour.uuid">
-    <th mat-header-cell *matHeaderCellDef class="text-center">
-      {{ hour.startTime }} ~ {{ hour.endTime }}
-    </th>
-    <td mat-cell *matCellDef="let row" class="text-center">
-      <button
-        mat-icon-button
-        [disabled]="disabled"
-        *ngIf="row[hour.uuid] !== undefined"
-        (click)="onHourClick(row, hour)"
-        type="button"
-      >
-        {{ row[hour.uuid] }}
-      </button>
-    </td>
-  </ng-container>
-
-  <tr mat-header-row *matHeaderRowDef="displayedCols"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedCols"></tr>
-</table>
