+++ src/environments/environment.prod.ts
@@ -7,22 +7,13 @@ export const environment = {
     projectId: 'tainan-public-witnessing',
     storageBucket: 'tainan-public-witnessing.appspot.com',
     messagingSenderId: '1087540310197',
-    appId: '1:1087540310197:web:30dfe6842e3d12af',
+    appId: '1:1087540310197:web:30dfe6842e3d12af'
   },
   UUID_NAMESPACE: '7b921192-c856-5152-8444-bb08b1efac9b',
-  TAINAN_PUBLIC_WITNESSING_PERMISSION_TOKEN:
-    '3c8ec6c4-7a3b-43a6-82af-0d243d6a7fc7',
-  DAY: [
-    'SUNDAY',
-    'MONDAY',
-    'TUESDAY',
-    'WEDNESDAY',
-    'THURSDAY',
-    'FRIDAY',
-    'SATURDAY',
-  ],
+  TAINAN_PUBLIC_WITNESSING_PERMISSION_TOKEN: '3c8ec6c4-7a3b-43a6-82af-0d243d6a7fc7',
+  DAY: ['SUNDAY', 'MONDAY', 'TUESDAY', 'WEDNESDAY', 'THURSDAY', 'FRIDAY', 'SATURDAY',],
   MOMENT_LOCALES: {
     ['en']: 'en',
     ['zh']: 'zh-tw',
-  },
+  }
 };
