+++ tsconfig.json
@@ -20,7 +20,6 @@
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
     "strictInjectionParameters": true,
-    "skipLibCheck": true,
-    "strictTemplates": false
+    "skipLibCheck": true
   }
 }
